+++ NEWS.md
@@ -1,5 +1,7 @@
 # mkin 0.9.49.10 (unreleased)
 
+- 'nlme.mmkin': An nlme method for mmkin row objects and an associated class with plot method
+
 - 'mean_degparms, nlme_data, nlme_function': Three new functions to facilitate building nlme models from mmkin row objects
 
 - 'endpoints': Don't return the SFORB list component if it's empty. This reduces distraction and complies with the documentation